refactor: particpantCount long -> int

+++ src/main/kotlin/com/yapp/lettie/api/timecapsule/service/dto/TimeCapsuleDetailDto.kt
@@ -8,7 +8,7 @@ data class TimeCapsuleDetailDto(
     val title: String,
     val subtitle: String?,
     val openAt: LocalDateTime,
-    val participantCount: Long,
+    val participantCount: Int,
     val likeCount: Long,
     val isLiked: Boolean,
     val status: TimeCapsuleStatus,
