+++ src/main/kotlin/com/yapp/lettie/domain/timecapsule/repository/TimeCapsuleUserRepository.kt
@@ -4,5 +4,5 @@ import com.yapp.lettie.domain.timecapsule.entity.TimeCapsuleUser
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface TimeCapsuleUserRepository : JpaRepository<TimeCapsuleUser, Long> {
-    fun countByTimeCapsuleId(capsuleId: Long): Long
+    fun countByTimeCapsuleId(capsuleId: Long): Int
 }
